feature: table handlers
- move table handlers to custom hook,
- move Pagination actions to separate component
