+++ src/component/Table/Table.tsx
@@ -1,6 +1,3 @@
-import * as React from 'react';
-import { useTheme } from '@mui/material/styles';
-import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
@@ -9,161 +6,46 @@ import TableFooter from '@mui/material/TableFooter';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
-import IconButton from '@mui/material/IconButton';
-import FirstPageIcon from '@mui/icons-material/FirstPage';
-import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
-import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
-import LastPageIcon from '@mui/icons-material/LastPage';
 import { TableSortLabel, TableHead } from '@mui/material';
+import { PaginationAction } from '../Pagination/Pagination';
+import { useHandleTable } from '../../feature/hook/useHandleTable';
 
-interface TablePaginationActionsProps {
-  count: number;
-  page: number;
-  rowsPerPage: number;
-  onPageChange: (
-    event: React.MouseEvent<HTMLButtonElement>,
-    newPage: number,
-  ) => void;
+interface CustomTableInterface {
+  fnHandleTableRowClick:(apiData: RowInterface) => void;
 }
-
-function TablePaginationActions(props: TablePaginationActionsProps) {
-  const theme = useTheme();
-  const { count, page, rowsPerPage, onPageChange } = props;
-
-  const handleFirstPageButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>,
-  ) => {
-    onPageChange(event, 0);
-  };
-
-  const handleBackButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    onPageChange(event, page - 1);
-  };
-
-  const handleNextButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    onPageChange(event, page + 1);
-  };
-
-  const handleLastPageButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    onPageChange(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-  };
-
-  return (
-    <Box sx={{ flexShrink: 0, ml: 2.5 }}>
-      <IconButton
-        onClick={handleFirstPageButtonClick}
-        disabled={page === 0}
-        aria-label="first page"
-      >
-        {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
-      </IconButton>
-      <IconButton
-        onClick={handleBackButtonClick}
-        disabled={page === 0}
-        aria-label="previous page"
-      >
-        {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
-      </IconButton>
-      <IconButton
-        onClick={handleNextButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="next page"
-      >
-        {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
-      </IconButton>
-      <IconButton
-        onClick={handleLastPageButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="last page"
-      >
-        {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
-      </IconButton>
-    </Box>
-  );
-}
-
-interface RowInterface {
-  id:number,
-  name:string,
-  calories:number,
-  fat:number
-}
-function createData(id:number, name: string, calories: number, fat: number) {
-  return { id, name, calories, fat };
-}
-
-const rows = [
-  createData(1,'Cupcake', 305, 3.7),
-  createData(2,'Donut', 452, 25.0),
-  createData(3,'Eclair', 262, 16.0),
-  createData(4,'Frozen yoghurt', 159, 6.0),
-  createData(5,'Gingerbread', 356, 16.0),
-  createData(6,'Honeycomb', 408, 3.2),
-  createData(7,'Ice cream sandwich', 237, 9.0),
-  createData(8,'Jelly Bean', 375, 0.0),
-  createData(9,'KitKat', 518, 26.0),
-  createData(0,'Lollipop', 392, 0.2),
-  createData(10,'Marshmallow', 318, 0),
-  createData(11,'Nougat', 360, 19.0),
-  createData(12,'Oreo', 437, 18.0),
-].sort((a, b) => (a.calories < b.calories ? -1 : 1));
-
-export const CustomTable=()=> {
-  const [page, setPage] = React.useState(0);
-  const [rowsPerPage, setRowsPerPage] = React.useState(5);
-  const [order, setOrder] = React.useState<'asc' | 'desc'>('asc');
-  const [orderBy, setOrderBy] = React.useState<keyof RowInterface>('id');
-  
-  const handleRequestSort = (event: React.MouseEvent<unknown>, property: keyof RowInterface) => {
-    const isAsc = orderBy === property && order === 'asc';
-    setOrder(isAsc ? 'desc' : 'asc');
-    setOrderBy(property);
-  };
-  
-  const createSortHandler = (property: keyof RowInterface) => (event: React.MouseEvent<unknown>) => {
-    handleRequestSort(event, property);
-  };
-  
-
-  // Avoid a layout jump when reaching the last page with empty rows.
-  const emptyRows =
-    page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
-
-  const handleChangePage = (
-    event: React.MouseEvent<HTMLButtonElement> | null,
-    newPage: number,
-  ) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (
-    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ) => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-    // rows.map((row)=>row.)
-  };
-
+export const CustomTable = ({fnHandleTableRowClick}: CustomTableInterface) => {
+  const { 
+    rowsPerPage,
+    initialData,
+    page,
+    emptyRows,
+    orderBy,
+    createSortHandler,
+    rows,
+    handleChangePage,
+    handleChangeRowsPerPage,
+    order
+  } = useHandleTable()
   return (
     <TableContainer component={Paper}>
       <Table sx={{ minWidth: 500 }} aria-label="custom pagination table">
         <TableBody>
           {(rowsPerPage > 0
-            ? rows.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-            : rows
-          ).map((row) => (
-            <TableRow key={row.name}>
+            ? initialData.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+            : initialData
+          ).map((initialData) => (
+            <TableRow onClick={()=>fnHandleTableRowClick(initialData)} key={initialData.name}>
                <TableCell component="th" scope="row">
-                {row.id.toString()}
+                {initialData.id.toString()}
               </TableCell>
-              <TableCell component="th" scope="row">
-                {row.name}
-              </TableCell>
-              <TableCell style={{ width: 160 }} align="right">
-                {row.calories}
+              <TableCell component="th" scope="initialData">
+                {initialData.name}
               </TableCell>
               <TableCell style={{ width: 160 }} align="right">
-                {row.fat}
+                {initialData.calories}
+              </TableCell>
+              <TableCell style={{ width: 160 }} align="right">
+                {initialData.fat}
               </TableCell>
             </TableRow>
           ))}
@@ -196,7 +78,7 @@ export const CustomTable=()=> {
               }}
               onPageChange={handleChangePage}
               onRowsPerPageChange={handleChangeRowsPerPage}
-              ActionsComponent={TablePaginationActions}
+              ActionsComponent={PaginationAction}
             />
           </TableRow>
         </TableFooter>
