+++ src/component/index.ts
@@ -1,10 +1,10 @@
-import {Pagination} from "./Pagination/Pagination";
+import {PaginationAction} from "./Pagination/Pagination";
 import { CustomTable } from "./Table/Table";
 import { CustomModal } from "./Modal/Modal";
 
 
 export {
-   Pagination,
+   PaginationAction,
    CustomTable,
    CustomModal
 }
