+++ src/types.d.ts
@@ -16,5 +16,13 @@ global {
          text: string;
        }
    }
+
+
+   interface RowInterface {
+      id:number,
+      name:string,
+      calories:number,
+      fat:number
+   }
 }
 export {}
